Scrape raw HTML from NYTimes

+++ scripts/get_articles.py
@@ -13,15 +13,14 @@ import os
 import time
 from datetime import datetime
 from itertools import product
+from urllib.error import HTTPError, URLError
 from urllib.parse import urlparse
 from urllib.request import urlopen
 
-import lxml
 import ptvsd
 import requests
 from bs4 import BeautifulSoup
 from docopt import docopt
-from goose3 import Goose
 from joblib import Parallel, delayed
 from posixpath import normpath
 from pymongo import MongoClient
@@ -65,9 +64,8 @@ def retrieve_articles(root_dir, year, month, db):
     db.scraping.insert_one({'year': year, 'month': month})
 
 
-def get_soup(url):
-    sauce = requests.get(url)
-    soup = BeautifulSoup(sauce.content, 'html.parser')
+def get_soup(raw_html):
+    soup = BeautifulSoup(raw_html, 'html.parser')
     [s.extract() for s in soup('script')]
     [s.extract() for s in soup('noscript')]
     figcap = soup.find_all("figcaption")
@@ -76,11 +74,13 @@ def get_soup(url):
 
 def retrieve_article(article, root_dir, db):
     result = db.articles.find_one({'_id': article['_id']})
-    if result is not None and result['scraped'] == True:
+    if result is not None and (result['scraped'] or result['error']):
         return
 
     data = article
     data['scraped'] = False
+    data['parsed'] = False
+    data['error'] = False
     data['pub_date'] = datetime.strptime(article['pub_date'],
                                          '%Y-%m-%dT%H:%M:%S%z')
 
@@ -92,35 +92,57 @@ def retrieve_article(article, root_dir, db):
 
     url = resolve_url(article['web_url'])
     while True:
+        #     try:
+        #         with Goose() as g:
+        #             extract = g.extract(url=url)
+        #             break
+        #     except requests.exceptions.MissingSchema:
+        #         return  # Ignore invalid URLs
+        #     except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectionError):
+        #         return
+        #     except requests.exceptions.TooManyRedirects:
+        #         return
+        #     except lxml.etree.ParserError:
+        #         return
+
         try:
-            with Goose() as g:
-                extract = g.extract(url=url)
-                break
-        except requests.exceptions.MissingSchema:
-            return  # Ignore invalid URLs
-        except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectionError):
+            response = urlopen(url)
+            break
+        except (ValueError, HTTPError):
+            # ValueError: unknown url type: '/interactive/2018/12/05/business/05Markets.html'
+            # urllib.error.HTTPError: HTTP Error 404: Not Found
+            data['error'] = True
+            db.articles.find_one_and_update(
+                {'_id': article['_id']}, {'$set': data})
             return
-        except requests.exceptions.TooManyRedirects:
-            return
-        except lxml.etree.ParserError:
-            return
+        except URLError:
+            # urllib.error.URLError: <urlopen error [Errno 110] Connection timed out>
+            time.sleep(60)
 
     data['web_url'] = url
-    data['article'] = extract.cleaned_text
+    raw_html = response.read().decode('utf-8')
+    data['raw_html'] = raw_html
+    # data['article'] = extract.cleaned_text
 
     if article['multimedia']:
         data['images'] = {}
-        _, figcap = get_soup(url)
+        _, figcap = get_soup(raw_html)
         figcap = [c for c in figcap if c.text]
         for ix, cap in enumerate(figcap):
             if cap.parent.attrs.get('itemid', 0):
+                img_path = os.path.join(root_dir, 'images',
+                                        f"{article['_id']}_{ix}.jpg")
+                if os.path.exists(img_path):
+                    text = cap.get_text().split('credit')[0]
+                    text = text.split('Credit')[0]
+                    data['images'].update({f'{ix}': text})
+                    continue
+
                 img_url = resolve_url(cap.parent.attrs['itemid'])
                 try:
                     img_data = requests.get(img_url, stream=True).content
                 except requests.exceptions.MissingSchema:
                     continue
-                img_path = os.path.join(root_dir, 'images',
-                                        f"{article['_id']}_{ix}.jpg")
 
                 with open(img_path, 'wb') as f:
                     f.write(img_data)
@@ -154,7 +176,7 @@ def main():
         ptvsd.enable_attach(address)
         ptvsd.wait_for_attach()
 
-    years = range(2018, 1979, -1)
+    years = range(2018, 1999, -1)
     months = range(12, 0, -1)
 
     root_dir = args['root_dir']
